+++ packages/mobile/shim.js
@@ -1,3 +1,0 @@
-if (typeof Buffer === 'undefined') {
-  global.Buffer = require('buffer').Buffer;
-}