+++ packages/mobile/src/store/init/startConnection/config.ts
@@ -1,7 +1,0 @@
-const socket = {
-  address: 'http://localhost:4677'
-}
-
-export default {
-  socket
-}